feat: logout component
