+++ src/modules/dashboard/layout/Sidebar.tsx
@@ -1,27 +1,13 @@
-"use client"
-
 import React from "react";
 import Link from "next/link";
 import DashboardIcon from "@/src/icons/DashboardIcon";
 import ProfileIcon from "@/src/icons/ProfileIcon";
 import TaskIcon from "@/src/icons/TaskIcon";
-import LogoutIcon from "@/src/icons/LogoutIcon";
-import { clearAuthCookie } from "@/app/actions/auth";
-import { useRouter } from "next/navigation";
-import { useDispatch } from "react-redux";
-import { logOut } from "@/src/redux/slices/authSlice";
-import toast from "react-hot-toast";
+import Logout from "../components/Logout";
+
 
 function DashboardSidebar() {
-  const dispatch = useDispatch();
-  const router = useRouter();
-
-  const handleLogout = async () => {
-    toast.success("خروج از پنل کاربری")
-    await clearAuthCookie();
-    dispatch(logOut());
-    router.push("/login");
-  };
+  
   return (
     <div className="hidden lg:block w-[280px] h-full bg-base-100 rounded-lg px-6 py-4 pb-6">
       <div className="w-full h-full flex flex-col items-start justify-between">
@@ -39,10 +25,7 @@ function DashboardSidebar() {
             وظایف
           </Link>
         </div>
-        <button className="logout-btn" onClick={handleLogout}>
-          <LogoutIcon />
-          خروج از حساب
-        </button>
+        <Logout/>
       </div>
     </div>
   );
